+++ src/ElementSetRoute/SetRouteFromParsed/SetRouteFirstRunCallback/CallbackForNoRoutingCallback.php
@@ -15,6 +15,8 @@ use function ltrim;
 
 class CallbackForNoRoutingCallback implements SetRouteFirstRunCallbackContract
 {
+    private const SPECIAL_URI_MAP = ['/__home__' => '/'];
+
     /**
      * @phpstan-ignore-next-line
      */
@@ -25,12 +27,19 @@ class CallbackForNoRoutingCallback implements SetRouteFirstRunCallbackContract
     ) {
     }
 
+    private function getFromMap(string $uri): string
+    {
+        return self::SPECIAL_URI_MAP[$uri] ?? $uri;
+    }
+
     public function call(App $app): void
     {
         $uri = '/' . ltrim(
             (string) $this->element->uri,
             '/',
         );
+
+        $uri = $this->getFromMap($uri);
 
         RouteParams::addParam(name: 'element', val: $this->element);